+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]]
 
+### Added
+- Support for `Token Escrow` (XLS-85d)
+
 ## [4.2.0] - 2025-6-09
 
 ### Added